Look for MSBuild in more places
SignTool.exe takes the path to msbuild.exe as an argument. Currently, we
use vswhere.exe to locate an appropriate VS installation, then look for
msbuild.exe at a location relative to that. In VS 2017, this relative
path was "MSBuild\15.0\Bin\msbuild.exe", but in VS 2019 that has
switched to "MSBuild\Current\Bin\msbuild.exe". Here we update Sign.proj
to look for both, preferring "Current" but using "15.0" if that is all
that is available.